requirements.txt
pillow~=10.3.0
numpy~=2.0.0